Move non-AppComponent code out from the AppComponent.kt file.
