+++ .idea/misc.xml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EntryPointsManager">
-    <list size="1">
+    <list size="3">
       <item index="0" class="java.lang.String" itemvalue="android.support.annotation.MainThread" />
+      <item index="1" class="java.lang.String" itemvalue="dagger.android.ContributesAndroidInjector" />
+      <item index="2" class="java.lang.String" itemvalue="dagger.android.ContributesAndroidInjector" />
     </list>
   </component>
   <component name="NullableNotNullManager">
