+++ app/src/main/java/com/aconno/sensorics/dagger/actionlist/ActionListFragmentsModule.kt
@@ -1,0 +1,13 @@
+package com.aconno.sensorics.dagger.actionlist
+
+import com.aconno.sensorics.dagger.application.FragmentScope
+import com.aconno.sensorics.ui.ActionListFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class ActionListFragmentsModule {
+    @FragmentScope
+    @ContributesAndroidInjector
+    abstract fun bindActionListFragment(): ActionListFragment
+}
