+++ app/src/main/java/com/aconno/sensorics/dagger/application/FragmentScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.application
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class FragmentScope
