+++ app/src/main/java/com/aconno/sensorics/dagger/deviceselect/DeviceSelectionFragmentsModule.kt
@@ -1,0 +1,12 @@
+package com.aconno.sensorics.dagger.deviceselect
+
+import com.aconno.sensorics.ui.settings.publishers.DeviceSelectFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class DeviceSelectionFragmentsModule {
+    @DeviceSelectScope
+    @ContributesAndroidInjector(modules = [DeviceSelectModule::class])
+    abstract fun bindDeviceSelectFragment(): DeviceSelectFragment
+}
