+++ app/src/main/java/com/aconno/sensorics/dagger/publish/PublishListFragmentsModule.kt
@@ -1,0 +1,13 @@
+package com.aconno.sensorics.dagger.publish
+
+import com.aconno.sensorics.dagger.application.FragmentScope
+import com.aconno.sensorics.ui.settings.publishers.PublishListFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class PublishListFragmentsModule {
+    @ContributesAndroidInjector
+    @FragmentScope
+    abstract fun bindPublishListFragment(): PublishListFragment
+}
